+++ ESP32-WROVER-Xtensa-FreeRTOS-bsp/Projects/microej/net/inc/LLNET_configuration.h
@@ -1,7 +1,7 @@
 /*
  * C
  *
- * Copyright 2017-2020 MicroEJ Corp. All rights reserved.
+ * Copyright 2017-2021 MicroEJ Corp. All rights reserved.
  * This library is provided in source code for use, modification and test, subject to license terms.
  * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
  * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
@@ -99,7 +99,7 @@
 /**
  * Enable network debug trace
  */
-//#define LLNET_DEBUG
+// #define LLNET_DEBUG
 
 /**
  * If ioctl() cannot be used to implement the LLNET_STREAMSOCKETCHANNEL_IMPL_available() function,