+++ ESP32-WROVER-Xtensa-FreeRTOS-bsp/Projects/microej/ssl/inc/LLNET_SSL_verifyCallback.h
@@ -1,14 +1,28 @@
 /*
  * C
  *
- * Copyright 2018-2020 MicroEJ Corp. All rights reserved.
+ * Copyright 2018-2021 MicroEJ Corp. All rights reserved.
  * This library is provided in source code for use, modification and test, subject to license terms.
  * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
  */
+
+/**
+ * @file
+ * @brief LLNET_SSL_verifyCallback functions for mbedtls.
+ * @author MicroEJ Developer Team
+ * @version 2.1.4
+ * @date 23 August 2021
+ */
+
 #ifndef LLNET_SSL_VERIFY_CALLBACK
 #define LLNET_SSL_VERIFY_CALLBACK
 
-#include "mbedtls/ssl.h"
+#if !defined(MBEDTLS_CONFIG_FILE)
+#include "mbedtls/config.h"
+#else
+#include MBEDTLS_CONFIG_FILE
+#endif
+#include "mbedtls/x509.h"
 
 #ifdef __cplusplus
 	extern "C" {
@@ -34,4 +48,4 @@ int LLNET_SSL_VERIFY_verifyCallback(void *data, mbedtls_x509_crt *crt, int depth
 	}
 #endif
 
-#endif
+#endif //LLNET_SSL_VERIFY_CALLBACK