+++ ESP32-WROVER-Xtensa-FreeRTOS-bsp/Projects/microej/ssl/src/LLNET_SSL_CONTEXT_impl.c
@@ -1,40 +1,55 @@
 /*
  * C
  *
- * Copyright 2018-2020 MicroEJ Corp. All rights reserved.
+ * Copyright 2018-2021 MicroEJ Corp. All rights reserved.
  * This library is provided in source code for use, modification and test, subject to license terms.
  * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
  */
 
-#include <LLNET_SSL_CONTEXT_impl.h>
-#include <LLNET_SSL_CONSTANTS.h>
-#include <net_ssl_errors.h>
-#include <LLNET_SSL_utils_mbedtls.h>
-#include "LLNET_SSL_verifyCallback.h"
-#include "microej_allocator.h"
-#include "ssl_parse.h"
-#include <stdlib.h>
+/**
+ * @file
+ * @brief LLNET_SSL_CONTEXT implementation over mbedtls.
+ * @author MicroEJ Developer Team
+ * @version 2.1.4
+ * @date 23 August 2021
+ */
 
+#if !defined(MBEDTLS_CONFIG_FILE)
+#include "mbedtls/config.h"
+#else
+#include MBEDTLS_CONFIG_FILE
+#endif
 #include "mbedtls/debug.h"
 #include "mbedtls/net.h"
 #include "mbedtls/ssl.h"
 #include "mbedtls/ssl_internal.h"
-#include "mbedtls/entropy.h"
-#include "mbedtls/ctr_drbg.h"
 #include "mbedtls/error.h"
 #include "mbedtls/platform.h"
-
-/* Check that the custom Mbedtls configuration file (MBEDTLS_CONFIG_FILE) is defined by the build definition*/
-#ifndef MBEDTLS_CONFIG_FILE
-#error "Custom configuration file (MBEDTLS_CONFIG_FILE) should be defined by the build definition"
+#if defined(MBEDTLS_ENTROPY_C) && defined(MBEDTLS_CTR_DRBG_C)
+#include "mbedtls/entropy.h"
+#include "mbedtls/ctr_drbg.h"
 #endif
+#include "LLNET_SSL_utils_mbedtls.h"
+#include "LLNET_SSL_verifyCallback.h"
+#include "LLNET_SSL_CONTEXT_impl.h"
+#include "LLNET_SSL_CONSTANTS.h"
+#include "LLNET_SSL_ERRORS.h"
+#include <stdlib.h>
+#include <string.h>
+
 
 #ifdef __cplusplus
 	extern "C" {
 #endif
 
 /* ----------- external function and variables ----------- */
+extern int32_t LLNET_SSL_TranslateReturnCode(int32_t mbedtls_error);
+#if defined(MBEDTLS_ENTROPY_C) && defined(MBEDTLS_CTR_DRBG_C)
 extern mbedtls_ctr_drbg_context ctr_drbg;
+#endif
+
+/* ----------- Definitions  -----------*/
+#define PEM_END "-----END CERTIFICATE-----"
 
 /* ----------- Private API  -----------*/
 #if MBEDTLS_DEBUG_LEVEL > 0
@@ -42,7 +57,7 @@ extern mbedtls_ctr_drbg_context ctr_drbg;
  * Debug callback for mbed TLS
  * Just prints on the USB serial port
  */
-void my_debug(void *ctx, int level, const char *file, int line,
+void microej_mbedtls_debug(void *ctx, int level, const char *file, int line,
 					 const char *str)
 {
 	const char *p, *basename;
@@ -64,9 +79,9 @@ void my_debug(void *ctx, int level, const char *file, int line,
 int32_t LLNET_SSL_CONTEXT_IMPL_createContext(int32_t protocol, uint8_t isClientContext, uint8_t retry){
 	LLNET_SSL_DEBUG_TRACE("%s(protocol=%d, isClientContext=%d, retry=%d)\n", __func__, (int)protocol, isClientContext, retry);
 
-	mbedtls_ssl_config* conf = (mbedtls_ssl_config*)microej_malloc(sizeof(mbedtls_ssl_config));
-
-	if (conf != NULL)
+	mbedtls_ssl_config* conf = (mbedtls_ssl_config*)mbedtls_calloc(1, sizeof(mbedtls_ssl_config));
+	void* p_rng = NULL;
+	if (NULL != conf)
 	{
 		mbedtls_ssl_config_init(conf);
 
@@ -85,38 +100,31 @@ int32_t LLNET_SSL_CONTEXT_IMPL_createContext(int32_t protocol, uint8_t isClientC
 		}
 
 		switch (protocol) {
-#ifdef MBEDTLS_SSL_PROTO_SSL3
+
 			case SSLv3_PROTOCOL:
 				mbedtls_ssl_conf_min_version(conf, MBEDTLS_SSL_MAJOR_VERSION_3, MBEDTLS_SSL_MINOR_VERSION_0 );
 				mbedtls_ssl_conf_max_version(conf, MBEDTLS_SSL_MAJOR_VERSION_3, MBEDTLS_SSL_MINOR_VERSION_0 );
 				break;
-#endif // MBEDTLS_SSL_PROTO_SSL3
-
-#ifdef MBEDTLS_SSL_PROTO_TLS1
 			case TLSv1_PROTOCOL:
 				mbedtls_ssl_conf_min_version(conf, MBEDTLS_SSL_MAJOR_VERSION_3, MBEDTLS_SSL_MINOR_VERSION_1 );
 				mbedtls_ssl_conf_max_version(conf, MBEDTLS_SSL_MAJOR_VERSION_3, MBEDTLS_SSL_MINOR_VERSION_1 );
 				break;
-#endif // MBEDTLS_SSL_PROTO_TLS1
-
-#ifdef MBEDTLS_SSL_PROTO_TLS1_1
 			case TLSv1_1_PROTOCOL:
 				mbedtls_ssl_conf_min_version(conf, MBEDTLS_SSL_MAJOR_VERSION_3, MBEDTLS_SSL_MINOR_VERSION_2 );
 				mbedtls_ssl_conf_max_version(conf, MBEDTLS_SSL_MAJOR_VERSION_3, MBEDTLS_SSL_MINOR_VERSION_2 );
 				break;
-#endif //MBEDTLS_SSL_PROTO_TLS1_1
-
-#ifdef MBEDTLS_SSL_PROTO_TLS1_2
 			case TLSv1_2_PROTOCOL:
 				mbedtls_ssl_conf_min_version(conf, MBEDTLS_SSL_MAJOR_VERSION_3, MBEDTLS_SSL_MINOR_VERSION_3 );
 				mbedtls_ssl_conf_max_version(conf, MBEDTLS_SSL_MAJOR_VERSION_3, MBEDTLS_SSL_MINOR_VERSION_3 );
 				break;
-#endif // MBEDTLS_SSL_PROTO_TLS1_2
 			default:
 				break;
 		}
 
-		mbedtls_ssl_conf_rng(conf, mbedtls_ctr_drbg_random, &ctr_drbg);
+#if defined(MBEDTLS_ENTROPY_C) && defined(MBEDTLS_CTR_DRBG_C)
+		p_rng = &ctr_drbg;
+#endif
+		mbedtls_ssl_conf_rng(conf, LLNET_SSL_utils_mbedtls_random, p_rng);
 
 		/* It is possible to disable authentication by passing
 		 * MBEDTLS_SSL_VERIFY_NONE in the call to mbedtls_ssl_conf_authmode()
@@ -126,9 +134,12 @@ int32_t LLNET_SSL_CONTEXT_IMPL_createContext(int32_t protocol, uint8_t isClientC
 		/*
 		 * Change certificate verification profile, allocate and initialize new profile
 		 */
-		mbedtls_x509_crt_profile* crt_profile = (mbedtls_x509_crt_profile*)microej_malloc(sizeof(mbedtls_x509_crt_profile));
-		if (crt_profile == NULL)
+		mbedtls_x509_crt_profile* crt_profile = (mbedtls_x509_crt_profile*)mbedtls_calloc(1, sizeof(mbedtls_x509_crt_profile));
+		if (NULL == crt_profile){
+			mbedtls_ssl_config_free(conf);
+			mbedtls_free(conf);
 			return J_CREATE_SSL_CONTEXT_ERROR;
+		}
 		/* Hashes from MD5 and above */
 		crt_profile->allowed_mds = MBEDTLS_X509_ID_FLAG( MBEDTLS_MD_MD5 ) |
 								   MBEDTLS_X509_ID_FLAG( MBEDTLS_MD_SHA1 ) |
@@ -147,9 +158,14 @@ int32_t LLNET_SSL_CONTEXT_IMPL_createContext(int32_t protocol, uint8_t isClientC
 		mbedtls_ssl_conf_cert_profile(conf, crt_profile);
 
 		/* Allocate and initialize certificate verify context*/
-		cert_verify_ctx* verify_ctx = (cert_verify_ctx*)microej_malloc(sizeof(cert_verify_ctx));
-		if (verify_ctx == NULL)
+		cert_verify_ctx* verify_ctx = (cert_verify_ctx*)mbedtls_calloc(1, sizeof(cert_verify_ctx));
+		if (NULL == verify_ctx){
+			mbedtls_ssl_conf_cert_profile(conf, NULL);
+			mbedtls_free(crt_profile);
+			mbedtls_ssl_config_free(conf);
+			mbedtls_free(conf);
 			return J_CREATE_SSL_CONTEXT_ERROR;
+		}
 
         verify_ctx->conf = conf;
         verify_ctx->isUnTrustCA = 0;
@@ -157,7 +173,7 @@ int32_t LLNET_SSL_CONTEXT_IMPL_createContext(int32_t protocol, uint8_t isClientC
 		mbedtls_ssl_conf_verify(conf, LLNET_SSL_VERIFY_verifyCallback, (void*)verify_ctx);
 
 #if MBEDTLS_DEBUG_LEVEL > 0
-			mbedtls_ssl_conf_dbg(conf, my_debug, NULL);
+			mbedtls_ssl_conf_dbg(conf, microej_mbedtls_debug, NULL);
 #if defined(MBEDTLS_DEBUG_C)
 			mbedtls_debug_set_threshold(MBEDTLS_DEBUG_LEVEL);
 #endif
@@ -178,17 +194,17 @@ int32_t LLNET_SSL_CONTEXT_IMPL_addTrustedCert(int32_t contextID, uint8_t *cert, 
 	int32_t ret = J_SSL_NO_ERROR;
 
 	/* Check parameters */
-	if (((uint8_t *)(cert + off) == NULL) || (len == 0) ||
-		((format != CERT_DER_FORMAT) && (format != CERT_PEM_FORMAT)))
+	if ((NULL == (uint8_t *)(cert + off)) || (0 == len) ||
+		((CERT_DER_FORMAT != format) && (CERT_PEM_FORMAT != format)))
 	{
 		return J_BAD_FUNC_ARG;
 	}
 
 	mbedtls_ssl_config* conf = (mbedtls_ssl_config*)(contextID);
 	mbedtls_x509_crt* cacert = NULL;
-	if (conf->ca_chain == NULL)
+	if (NULL == conf->ca_chain)
 	{
-		cacert = (mbedtls_x509_crt*)microej_malloc(sizeof(mbedtls_x509_crt));
+		cacert = (mbedtls_x509_crt*)mbedtls_calloc(1, sizeof(mbedtls_x509_crt));
 		mbedtls_x509_crt_init(cacert);
 	}
 	else
@@ -196,21 +212,31 @@ int32_t LLNET_SSL_CONTEXT_IMPL_addTrustedCert(int32_t contextID, uint8_t *cert, 
 		cacert = conf->ca_chain;
 	}
 
-	if (conf != NULL && cacert != NULL)
+	if (NULL != conf && NULL != cacert)
 	{
 		int ret;
 
-		if (format == CERT_DER_FORMAT) {
+		if (CERT_DER_FORMAT == format) {
 			ret = mbedtls_x509_crt_parse_der(cacert, (const unsigned char *) (cert + off), len);
+            ret = LLNET_SSL_TranslateReturnCode(ret);
 		} else {
-			ret = microej_x509_crt_parse(cacert, cert, off, len);
+			/* Find end of first certificate. */
+			char* end = strstr((char*) cert, PEM_END);
+			if (NULL == end) {
+				ret = J_CREATE_SSL_CONTEXT_ERROR;
+			} else {
+				end += sizeof(PEM_END);
+				int end_index = end - (char*) cert;
+				ret = LLNET_SSL_utils_mbedtls_x509_crt_parse(cacert, (uint8_t *) cert, off, end_index);
+			}
 		}
 
-		if (ret != 0) {
-			return NET_SSL_TranslateReturnCode(ret);
+		if(J_SSL_NO_ERROR != ret){
+			mbedtls_free(cacert);
+			return ret;
 		}
 
-		if (conf->ca_chain == NULL){
+		if (NULL == conf->ca_chain){
 			mbedtls_ssl_conf_ca_chain(conf, cacert, NULL);
 		}
 	}
@@ -228,14 +254,14 @@ int32_t LLNET_SSL_CONTEXT_IMPL_clearTrustStore(int32_t contextID, uint8_t retry)
 	LLNET_SSL_DEBUG_TRACE("%s(context=%d, retry=%d)\n", __func__, (int)contextID, retry);
 	mbedtls_ssl_config* conf = (mbedtls_ssl_config*)(contextID);
 
-	if (conf != NULL)
+	if (NULL != conf)
 	{
-		if (conf->ca_chain != NULL)
+		if (NULL != conf->ca_chain)
 		{
 			void* chain_ptr = (void*)conf->ca_chain;
 			mbedtls_ssl_conf_ca_chain(conf, NULL, NULL);
 			mbedtls_x509_crt_free(chain_ptr);
-			microej_free(chain_ptr);
+			mbedtls_free(chain_ptr);
 		}
 	}
 
@@ -247,14 +273,14 @@ int32_t LLNET_SSL_CONTEXT_IMPL_clearKeyStore(int32_t contextID, uint8_t retry){
 	mbedtls_ssl_config* conf = (mbedtls_ssl_config*)(contextID);
 
 	/* Free the private key */
-	if (conf != NULL)
+	if (NULL != conf)
 	{
-		if (conf->key_cert != NULL)
+		if (NULL != conf->key_cert)
 		{
-			if (conf->key_cert->key != NULL)
+			if (NULL != conf->key_cert->key)
 			{
 				mbedtls_pk_free(conf->key_cert->key);
-				microej_free(conf->key_cert->key);
+				mbedtls_free(conf->key_cert->key);
 			}
 		}
 	}
@@ -265,42 +291,50 @@ int32_t LLNET_SSL_CONTEXT_IMPL_clearKeyStore(int32_t contextID, uint8_t retry){
 int32_t LLNET_SSL_CONTEXT_IMPL_setCertificate(int32_t contextID, uint8_t* cert, int32_t offset, int32_t len, int32_t format, uint8_t retry){
 	LLNET_SSL_DEBUG_TRACE("%s(context=%d, retry=%d)\n", __func__, (int)contextID, retry);
 	mbedtls_ssl_config* conf = (mbedtls_ssl_config*)(contextID);
-	mbedtls_x509_crt *clicert = NULL;
+	mbedtls_x509_crt *clicert;
 	int ret;
 
 	/* Check parameters */
-	if (((uint8_t *)((int32_t)cert + offset) == NULL) || (len == 0) ||
-		((format != CERT_DER_FORMAT) && (format != CERT_PEM_FORMAT))) {
+	if ((NULL == (uint8_t *)((int32_t)cert + offset)) || (0 == len) ||
+		((CERT_DER_FORMAT != format) && (CERT_PEM_FORMAT != format))) {
 		return J_BAD_FUNC_ARG;
 	}
 
 	/* Allocate a new keycert if needed, otherwise free the existing certificate */
-	if (conf->key_cert == NULL) {
-		conf->key_cert = microej_calloc( 1, sizeof( mbedtls_ssl_key_cert ) );
+	if (NULL == conf->key_cert) {
+		conf->key_cert = mbedtls_calloc( 1, sizeof( mbedtls_ssl_key_cert ) );
 	} else {
 		mbedtls_x509_crt_free(conf->key_cert->cert);
-		microej_free(conf->key_cert->cert);
+		mbedtls_free(conf->key_cert->cert);
+	}
+
+	if (NULL == conf->key_cert){
+		return J_MEMORY_ERROR;
 	}
 
 	/* The LLNET_SSL_* interface does not support multiple keycert (only one private key can be set) */
 	conf->key_cert->next = NULL;
 
 	/* Allocate a new certificate */
-	clicert = microej_calloc( 1, sizeof( mbedtls_x509_crt ) );
+	clicert = mbedtls_calloc( 1, sizeof( mbedtls_x509_crt ) );
+	if (NULL == clicert){
+		return J_MEMORY_ERROR;
+	}
 
 	/* Try to parse the certificate */
 	mbedtls_x509_crt_init(clicert);
-	if (format == CERT_DER_FORMAT) {
+	if (CERT_DER_FORMAT == format) {
 		ret = mbedtls_x509_crt_parse_der(clicert, (uint8_t *) ((int32_t)cert + offset), (uint32_t) len);
+		ret = LLNET_SSL_TranslateReturnCode(ret);
 	} else {
-			ret = microej_x509_crt_parse(clicert, cert, offset, len);
+		ret = LLNET_SSL_utils_mbedtls_x509_crt_parse(clicert, (uint8_t *) cert, offset, len);
 	}
 
 	/* Check parse result */
-	if (ret != 0) {
+	if (J_SSL_NO_ERROR != ret) {
 		/* Free the certificate */
 		mbedtls_x509_crt_free(clicert);
-		microej_free(clicert);
+		mbedtls_free(clicert);
 
 		return J_CERT_PARSE_ERROR;
 	}
@@ -319,7 +353,7 @@ int32_t LLNET_SSL_CONTEXT_IMPL_setPrivateKey(int32_t contextID, uint8_t* private
 	uint8_t * pwd = NULL;
 
 	/* Check parameters */
-	if (((uint8_t *)((int32_t)privateKey + offset) == NULL) || (len == 0)) {
+	if ((NULL == (uint8_t *)((int32_t)privateKey + offset)) || (0 == len)) {
 		return J_BAD_FUNC_ARG;
 	}
 
@@ -329,30 +363,37 @@ int32_t LLNET_SSL_CONTEXT_IMPL_setPrivateKey(int32_t contextID, uint8_t* private
 	}
 
 	/* Allocate a new keycert if needed, otherwise free the existing private key */
-	if (conf->key_cert == NULL) {
-		conf->key_cert = microej_calloc( 1, sizeof( mbedtls_ssl_key_cert ) );
+	if (NULL == conf->key_cert) {
+		conf->key_cert = mbedtls_calloc( 1, sizeof( mbedtls_ssl_key_cert ) );
 	} else {
 		mbedtls_pk_free(conf->key_cert->key);
-		microej_free(conf->key_cert->key);
+		mbedtls_free(conf->key_cert->key);
+	}
+
+	if (NULL == conf->key_cert){
+		return J_MEMORY_ERROR;
 	}
 
 	/* The LLNET_SSL_* interface does not support multiple keycert (only one private key can be set) */
 	conf->key_cert->next = NULL;
 
 	/* Allocate a new private key */
-	key = microej_calloc( 1, sizeof( mbedtls_pk_context ) );
+	key = mbedtls_calloc( 1, sizeof( mbedtls_pk_context ) );
+	if (NULL == key){
+		return J_MEMORY_ERROR;
+	}
 
 	/* Try to parse the private key */
 	mbedtls_pk_init(key);
 	int mbed_err = mbedtls_pk_parse_key(key,
 			(uint8_t *) ((int32_t)privateKey + offset), (uint32_t) len,
 			pwd, (uint32_t) passwordLen);
-	if (mbed_err != 0) {
+	if (0 != mbed_err) {
 
 		print_mbedtls_error(__func__, mbed_err);
 		/* Free the private key */
 		mbedtls_pk_free(key);
-		microej_free(key);
+		mbedtls_free(key);
 
 		return J_NO_PRIVATE_KEY;
 	}
@@ -375,20 +416,21 @@ int32_t LLNET_SSL_CONTEXT_IMPL_addChainCertificate(int32_t contextID, uint8_t* c
 	int ret;
 
 	/* Check parameters */
-	if (((uint8_t *)((int32_t)cert + offset) == NULL) || (len == 0) ||
-		((format != CERT_DER_FORMAT) && (format != CERT_PEM_FORMAT))) {
+	if ((NULL == (uint8_t *)((int32_t)cert + offset)) || (0 == len) ||
+		((CERT_DER_FORMAT != format) && (CERT_PEM_FORMAT != format))) {
 		return J_BAD_FUNC_ARG;
 	}
 
 	/* Try to parse the certificate, adding it to the chained list of certificated from keycert */
-	if (format == CERT_DER_FORMAT) {
+	if (CERT_DER_FORMAT == format) {
 		ret = mbedtls_x509_crt_parse_der(conf->key_cert->cert, (uint8_t *) ((int32_t)cert + offset), (uint32_t) len);
+		ret = LLNET_SSL_TranslateReturnCode(ret);
 	} else {
-		ret = microej_x509_crt_parse(conf->key_cert->cert, cert, offset, len);
+		ret = LLNET_SSL_utils_mbedtls_x509_crt_parse(conf->key_cert->cert, (uint8_t *) cert, offset, len);
 	}
 
 	/* Check parse result */
-	if (ret != 0) {
+	if (J_SSL_NO_ERROR != ret) {
 		/* Free the certificate */
 		mbedtls_x509_crt_free(conf->key_cert->cert);
 
@@ -402,47 +444,47 @@ int32_t LLNET_SSL_CONTEXT_IMPL_closeContext(int32_t contextID, uint8_t retry){
 	LLNET_SSL_DEBUG_TRACE("%s(context=%d, retry=%d)\n", __func__, (int)contextID, retry);
 	mbedtls_ssl_config* conf = (mbedtls_ssl_config*)(contextID);
 
-	if (conf != NULL)
+	if (NULL != conf)
 	{
-		if (conf->ca_chain != NULL)
+		if (NULL != conf->ca_chain)
 		{
 			void* chain_ptr = (void*)conf->ca_chain;
 			mbedtls_ssl_conf_ca_chain(conf, NULL, NULL);
 			mbedtls_x509_crt_free(chain_ptr);
-			microej_free(chain_ptr);
+			mbedtls_free(chain_ptr);
 		}
 
-		if (conf->p_vrfy != NULL)
+		if (NULL != conf->p_vrfy)
 		{
 			void* vrfy_ptr = (void*)conf->p_vrfy;
 			mbedtls_ssl_conf_verify(conf, NULL, NULL);
-			microej_free(vrfy_ptr);
+			mbedtls_free(vrfy_ptr);
 		}
 
-		if (conf->cert_profile != NULL)
+		if (NULL != conf->cert_profile)
 		{
 			void* profile_ptr = (void*)conf->cert_profile;
 			mbedtls_ssl_conf_cert_profile(conf, NULL);
-			microej_free(profile_ptr);
+			mbedtls_free(profile_ptr);
 		}
 
-		if (conf->key_cert != NULL)
+		if (NULL != conf->key_cert)
 		{
-			if (conf->key_cert->key != NULL)
+			if (NULL != conf->key_cert->key)
 			{
 				mbedtls_pk_free(conf->key_cert->key);
-				microej_free(conf->key_cert->key);
+				mbedtls_free(conf->key_cert->key);
 			}
 
-			if (conf->key_cert->cert != NULL)
+			if (NULL != conf->key_cert->cert)
 			{
 				mbedtls_x509_crt_free(conf->key_cert->cert);
-				microej_free(conf->key_cert->cert);
+				mbedtls_free(conf->key_cert->cert);
 			}
 		}
 
 		mbedtls_ssl_config_free(conf);
-		microej_free((void*)conf);
+		mbedtls_free((void*)conf);
 	}
 
 	return J_SSL_NO_ERROR;